refactor(controller): 提取 controller 层，七层单向依赖架构
- ui/controllers/ → controller/（独立第七层）
- 相对导入改为绝对导入
- UI 层 import 路径全部更新
- 依赖链：types → config → repo → service → runtime → controller → ui
