+++ src/danmaku_sender/controller/editor_controller.py
