+++ src/danmaku_sender/ui/editor/page.py
@@ -13,7 +13,7 @@ from .components import EditorTableModel, ValidationRulesGroup, PropertyInspecto
 from .dialogs import EditDanmakuDialog, TimeOffsetDialog, ArrayGeneratorDialog
 
 from danmaku_sender.ui.framework.style_loader import get_svg_icon
-from danmaku_sender.ui.controllers.editor_controller import EditorController
+from danmaku_sender.controller.editor_controller import EditorController
 from danmaku_sender.types.models.editor_types import EditorField, InsertPosition
 from danmaku_sender.runtime.app_state import AppState
