+++ src/danmaku_sender/ui/history/page.py
@@ -13,8 +13,8 @@ from .dialogs import DanmakuDetailDialog
 from danmaku_sender.types.models.common import DanmakuStatus
 from danmaku_sender.types.models.video import VideoInfo
 from danmaku_sender.runtime.app_state import AppState
-from danmaku_sender.ui.controllers.video_controller import VideoController
-from danmaku_sender.ui.controllers.history_controller import HistoryController
+from danmaku_sender.controller.video_controller import VideoController
+from danmaku_sender.controller.history_controller import HistoryController
 
 
 logger = logging.getLogger("App.System.UI.History")
