+++ src/danmaku_sender/ui/sender_page.py
@@ -12,8 +12,8 @@ from PySide6.QtCore import Qt, QDateTime, Signal, Slot
 
 from .framework.binder import UIBinder
 from .framework.style_loader import get_svg_icon
-from .controllers.video_controller import VideoController
-from .controllers.sender_controller import SenderController
+from danmaku_sender.controller.video_controller import VideoController
+from danmaku_sender.controller.sender_controller import SenderController
 
 from danmaku_sender.service.sender import SendingContext
 from danmaku_sender.types.models.video import VideoInfo
